fix: code cleaning import file modal
