+++ apps/bilan-carbone/src/components/base/ImportFileModal/WarningList.tsx
@@ -1,0 +1,38 @@
+import { ImportWarning } from '@/types/import.types'
+import { Alert, AlertTitle, List, ListItem, Typography } from '@mui/material'
+import { useTranslations } from 'next-intl'
+import styles from './ImportFileModal.module.css'
+import WarningItem from './WarningItem'
+
+interface Props {
+  warnings: { line: number; items: ImportWarning[] }[]
+  t: ReturnType<typeof useTranslations>
+  tCommon: ReturnType<typeof useTranslations>
+}
+
+const WarningList = ({ warnings, t, tCommon }: Props) => (
+  <Alert severity="warning">
+    <AlertTitle>{t('warningTitle')}</AlertTitle>
+    <List dense className={styles.errorList}>
+      {warnings.map(({ line, items }) => (
+        <ListItem key={line} disableGutters className="py025">
+          <div>
+            {line > 0 && (
+              <Typography variant="body2" fontWeight="medium">
+                {tCommon('label.line', { line })}
+                {items[0]?.sourceName ? ` — ${items[0].sourceName}` : ''}
+              </Typography>
+            )}
+            <List dense disablePadding>
+              {items.map((w, i) => (
+                <WarningItem key={i} w={w} line={line} t={t} />
+              ))}
+            </List>
+          </div>
+        </ListItem>
+      ))}
+    </List>
+  </Alert>
+)
+
+export default WarningList
